solve duplicated files
