--- src/component-overview/Overview.js
@@ -1,11 +1,0 @@
-import React, {Component} from "react";
-
-class Overview extends Component {
-  render() {
-    return (
-        <h2>Overview in progress...</h2>
-    );
-  }
-}
-
-export default Overview;
