+++ src/components/component-charts/ChartsUsage.js
@@ -1,31 +1,79 @@
 import React, { Component } from "react";
 import { Link } from 'react-router-dom';
-import { requestCharts } from '../../services/ChartsService';
+import { requestCharts } from './../../services/ChartsService';
 import TableCharts from './TableCharts';
 
 class ChartsUsage extends Component {
     constructor(props) {
-      super(props)
-    
-      this.state = {
-         chartData:[]
-      }
+        super(props)
+
+        this.state = {
+            chartData: [],
+            chartNames: []
+        }
+
+        this.renderChartData = this.renderChartData.bind(this);
+        this.renderTimesUsed = this.renderTimesUsed.bind(this);
+        this.renderTimesPercentage = this.renderTimesPercentage.bind(this);
+        this.renderChartUsers = this.renderChartUsers.bind(this);
     }
-    componentDidMount(){
-        this.fetchCharts()
+
+    componentDidMount() {
+        this.fetchCharts();
     }
 
     fetchCharts() {
         requestCharts()
             .then(data => {
                 this.setState({
-                    chartData: data.open_chart_events
+                    chartData: data.open_chart_events,
                 });
+
+                this.renderChartData(data.open_chart_events);
             });
     }
 
+    renderChartData(chartData) {
+        const mappedChartData = chartData.map(chart => {
+            return chart.details.chart_name;
+        });
+
+        this.setState({
+            chartNames: [...new Set(mappedChartData)]
+        })
+    }
+
+    renderTimesUsed(chart) {
+        const reducedChartData = this.state.chartData.reduce((acc, item) => {
+            if (item.details.chart_name === chart) {
+                acc++
+            }
+            return acc
+        }, 0);
+
+        return reducedChartData;
+    }
+
+    renderTimesPercentage(timesUsed) {
+        const timesPercentage = (timesUsed / this.state.chartData.length * 100).toFixed(1);
+        return timesPercentage;
+    }
+
+    renderChartUsers(givenChart) {
+        const originalCharts = this.state.chartData;
+
+        const mappedUsersData = originalCharts
+            .filter(chart => chart.details.chart_name === givenChart)
+            .map(chart => {
+                return chart.request.user__username;
+            });
+
+        return [...new Set(mappedUsersData)].length;
+    }
 
     render() {
+        const { chartNames } = this.state;
+
         return (
             <div className="app__container">
                 <main className="app__main">
@@ -41,13 +89,13 @@ class ChartsUsage extends Component {
                     </div>
                     <div className="charts__container">
                         <div className="table__container">
-                            <TableCharts chartData={this.state.chartData}/>
+                            <TableCharts chartNames={chartNames} renderTimesUsed={this.renderTimesUsed} renderTimesPercentage={this.renderTimesPercentage} renderChartUsers={this.renderChartUsers}/>
                         </div>
                         <div className="chart__filters">
                             <div className="chart__filters-options">
-                                    <h3>OPTIONS</h3>
-                                    <label>
-                                        <input type="radio"></input> exclude support users (x@stylesage.com)
+                                <h3>OPTIONS</h3>
+                                <label>
+                                    <input type="radio"></input> exclude support users (x@stylesage.com)
                                     </label>
                             </div>
                             <div className="chart__filters-range">
@@ -63,24 +111,22 @@ class ChartsUsage extends Component {
                                         <input type="radio" />last month
                                     </label>
                                 </div>
-                               <div>
+                                <div>
                                     <label>
-                                        <input type="radio"/>last 2 months
+                                        <input type="radio" />last 2 months
                                     </label>
-                               </div>
-                               <div>
+                                </div>
+                                <div>
                                     <label>
                                         <input type="radio" /> set date
-                                    </label> 
-                               </div>
+                                    </label>
+                                </div>
                                 <div>
                                     <label>
                                         <input type="radio" /> always
                                     </label>
                                 </div>
-                                
                             </div>
-                                    
                             <div className="chart__filters-groups">
                                 <h3>USER GROUPS</h3>
                                 <p>select all | select active | clear all</p>
@@ -93,8 +139,6 @@ class ChartsUsage extends Component {
                             </div>
                         </div>
                     </div>
-                  
-                    
                 </main>
             </div>
         );
